+++ hyperfit/src/main/java/org/hyperfit/Pipeline.java
@@ -3,8 +3,8 @@ package org.hyperfit;
 public interface Pipeline<I,O> {
 
     interface Step<I,O> {
-        O run(I request, Pipeline<I,O> pipeline);
+        O run(I input, Pipeline<I,O> pipeline);
     }
 
-    O run(I request);
+    O run(I input);
 }